Не нужен xr_vector в виде отдельного класса

+++ ogsr_engine/Layers/xrRender/r__dsgraph_render_lods.cpp
@@ -36,7 +36,7 @@ void R_dsgraph_structure::r_dsgraph_render_lods	(bool _setup_zb, bool _clear)
 
 	for (u32 i=0; i<lstLODs.size(); i++)
 	{
-		const u32	iBatchSize	= _min( lstLODs.size()-i, uiImpostersFit);
+		const u32	iBatchSize	= _min( u32(lstLODs.size())-i, uiImpostersFit);
 		int			cur_count	= 0;
 		u32			vOffset;
 		FLOD::_hw*	V			= (FLOD::_hw*)RCache.Vertex.Lock	(iBatchSize*uiVertexPerImposter,firstV->geom->vb_stride, vOffset);
